feat : add bookmark save

+++ ccs/user-api/src/main/java/zerobase/bud/news/dto/BookMarkDto.java
@@ -14,7 +14,7 @@ public class BookMarkDto {
 
     private long id;
     private long news_id;
-    private int user_id;
+    private long user_id;
 
 
     public static BookMarkDto fromEntity(BookMark bookMark){
